+++ packages/proxy/src/types.ts
@@ -36,7 +36,3 @@ export interface RouteResult {
   // the phase that this route is defined in
   phase?: HandleValue | null;
 }
-
-export interface ApiGatewayOriginProps {
-  path: string;
-}
